api/requirements.txt
aiohttp
blinker
brotlipy
cached-property
flask-cors
opencv-python
pip-chill
pyjwt
pyopenssl
pysocks
pyu2f
scipy
tensorflow-cpu
gunicorn
Pillow